healthcheck: refactor and create enough of FakeHealthCheck to allow unit testing of ScatterConn
