+++ go/vt/vtgate/tabletgateway.go
@@ -17,6 +17,8 @@ limitations under the License.
 package vtgate
 
 import (
+	"context"
+	"flag"
 	"fmt"
 	"math/rand"
 	"sort"
@@ -24,8 +26,6 @@ import (
 	"time"
 
 	"vitess.io/vitess/go/vt/topo/topoproto"
-
-	"golang.org/x/net/context"
 
 	"vitess.io/vitess/go/vt/discovery"
 	"vitess.io/vitess/go/vt/log"
@@ -48,42 +48,17 @@ func init() {
 	RegisterGatewayCreator(tabletGatewayImplementation, createTabletGateway)
 }
 
-//HealthCheck declares what the TabletGateway needs from the HealthCheck
-type HealthCheck interface {
-	// CacheStatus returns a displayable version of the health check cache.
-	CacheStatus() discovery.TabletsCacheStatusList
-
-	// Close stops the healthcheck.
-	Close() error
-
-	// WaitForAllServingTablets waits for at least one healthy serving tablet in
-	// each given target before returning.
-	// It will return ctx.Err() if the context is canceled.
-	// It will return an error if it can't read the necessary topology records.
-	WaitForAllServingTablets(ctx context.Context, targets []*querypb.Target) error
-
-	// TabletConnection returns the TabletConn of the given tablet.
-	TabletConnection(alias *topodatapb.TabletAlias) (queryservice.QueryService, error)
-
-	// RegisterStats registers the connection counts stats
-	RegisterStats()
-
-	// GetHealthyTabletStats returns only the healthy tablets.
-	// The returned array is owned by the caller.
-	// For TabletType_MASTER, this will only return at most one entry,
-	// the most recent tablet of type master.
-	// This returns a copy of the data so that callers can access without
-	// synchronization
-	GetHealthyTabletStats(target *querypb.Target) []*discovery.TabletHealth
-}
-
-var _ HealthCheck = (*discovery.HealthCheckImpl)(nil)
+var _ discovery.HealthCheck = (*discovery.HealthCheckImpl)(nil)
+var (
+	// CellsToWatch is the list of cells the healthcheck operates over. If it is empty, only the local cell is watched
+	CellsToWatch = flag.String("cells_to_watch", "", "comma-separated list of cells for watching tablets")
+)
 
 // TabletGateway implements the Gateway interface.
 // This implementation uses the new healthcheck module.
 type TabletGateway struct {
 	queryservice.QueryService
-	hc            HealthCheck
+	hc            discovery.HealthCheck
 	srvTopoServer srvtopo.Server
 	localCell     string
 	retryCount    int
@@ -99,11 +74,16 @@ type TabletGateway struct {
 }
 
 func createTabletGateway(ctx context.Context, _ discovery.LegacyHealthCheck, serv srvtopo.Server, cell string, _ int) Gateway {
-	return NewTabletGateway(ctx, serv, cell)
+	// we ignore the passed in LegacyHealthCheck and let TabletGateway create it's own HealthCheck
+	return NewTabletGateway(ctx, nil /*discovery.Healthcheck*/, serv, cell)
+}
+
+func createHealthCheck(ctx context.Context, retryDelay, timeout time.Duration, ts *topo.Server, cell, cellsToWatch string) discovery.HealthCheck {
+	return discovery.NewHealthCheck(ctx, retryDelay, timeout, ts, cell, cellsToWatch)
 }
 
 // NewTabletGateway creates and returns a new TabletGateway
-func NewTabletGateway(ctx context.Context, serv srvtopo.Server, localCell string) *TabletGateway {
+func NewTabletGateway(ctx context.Context, hc discovery.HealthCheck, serv srvtopo.Server, localCell string) *TabletGateway {
 	var topoServer *topo.Server
 	if serv != nil {
 		var err error
@@ -112,7 +92,9 @@ func NewTabletGateway(ctx context.Context, serv srvtopo.Server, localCell string
 			log.Exitf("Unable to create new TabletGateway: %v", err)
 		}
 	}
-	hc := discovery.NewHealthCheck(ctx, *HealthCheckRetryDelay, *HealthCheckTimeout, topoServer, localCell)
+	if hc == nil {
+		hc = createHealthCheck(ctx, *HealthCheckRetryDelay, *HealthCheckTimeout, topoServer, localCell, *CellsToWatch)
+	}
 
 	gw := &TabletGateway{
 		hc:                hc,
